% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bad_apples.R
\name{bad_apples}
\alias{bad_apples}
\title{bad_apples Documentation}
\usage{
bad_apples(a, b)
}
\arguments{
\item{b}{}
}
\value{

}
\description{
Example Description: Returns the names of numeric variables with outliers,
how many outliers are present, and which rows they are in.
}
