% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_ingredients.R
\name{find_missing_ingredients}
\alias{find_missing_ingredients}
\title{find_missing_ingredients Documentation}
\usage{
find_missing_ingredients(a, b)
}
\arguments{
\item{b}{}
}
\value{

}
\description{
Example Description: Returns the indices of rows with columns which have missing values,
as well as how many missing values they have
}
