% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruits_and_veg.R
\name{fruits_and_veg}
\alias{fruits_and_veg}
\title{fruits_and_veg Documentation}
\usage{
fruits_and_veg(a, b)
}
\arguments{
\item{b}{}
}
\value{

}
\description{
Example Description: Returns the names of numeric and categorical variables in the dataset
}
