% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_recipe.R
\name{make_recipe}
\alias{make_recipe}
\title{make_recipe}
\usage{
make_recipe(
  data,
  recipe,
  create_train_test = FALSE,
  create_train_valid_test = TRUE
)
}
\arguments{
\item{data}{A dataframe containing training data, validation data, X, and y}

\item{recipe}{A string specifying which recipe to apply to the data}

\item{create_train_test}{Boolean, if TRUE will partition data into train and test}

\item{create_train_valid_test}{Boolean, if TRUE will partition data into train, valid, and test}
}
\value{
A list of dataframes e.g. list(train, valid, test)
}
\description{
The `make_recipe()` function is used to quickly apply common data preprocessing techniques
}
