% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_ingredients.R
\name{missing_ingredients}
\alias{missing_ingredients}
\title{missing_ingredients Documentation}
\usage{
missing_ingredients(a, b)
}
\arguments{
\item{b}{}
}
\value{

}
\description{
Example Description: Returns the indices of rows with columns which have missing values,
as well as how many missing values they have
}
