% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_recipe.R
\name{preprocess_recipe}
\alias{preprocess_recipe}
\title{preprocess_recipe Documentation}
\usage{
preprocess_recipe(a, b)
}
\arguments{
\item{b}{}
}
\value{

}
\description{
Example Description: Preprocesses data...
}
