% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_and_dice.R
\name{slice_and_dice}
\alias{slice_and_dice}
\title{slice_and_dice Documentation}
\usage{
slice_and_dice(a, b)
}
\arguments{
\item{b}{}
}
\value{

}
\description{
Example Description: Preprocesses data...
}
