% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bad_apples.R
\name{find_bad_apples}
\alias{find_bad_apples}
\title{find_bad_apples}
\usage{
find_bad_apples(df)
}
\arguments{
\item{df}{A dataframe containing numeric data}
}
\value{
A dataframe with column for 'indices' and column for 'number of outliers'
}
\description{
This function uses a univariate approach to outlier detection.
It returns the indices of rows within a dataframe that contain variable values
which are at least 3 standard deviations from the norm of the variable (outliers),
as well as how many outliers exist within those rows.

Note: This function works best for small datasets with unimodal variable distributions.
}
