% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruits_and_veg.R
\name{find_fruits_veg}
\alias{find_fruits_veg}
\title{fruits_and_veg Documentation}
\usage{
find_fruits_veg(df, type_of_out = "categ")
}
\arguments{
\item{df}{an input dataframe}

\item{type_of_out}{a character indicats the wanted type (numeric/ categorical) of data}
}
\value{
a vector of index of wanted data type (numeric/ categorical)
}
\description{
Example Description: Returns the names of numeric and categorical variables in the dataset
}
\examples{
fruits_and_veg(df_demo, 'categ')
>>>> c(1)
}
