% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_ingredients.R
\name{find_missing_ingredients}
\alias{find_missing_ingredients}
\title{find_missing_ingredients Documentation}
\usage{
find_missing_ingredients(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe containing the columns: column_name, list_of_indices, na_count, na_proportion
}
\description{
For each column with missing values, this function will create a reference
list of row indices, sum the number and calculate proportion of missing values
}
\examples{
missing_ingredients(df)

}
