% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_ingredients.R
\name{find_missing_ingredients}
\alias{find_missing_ingredients}
\title{find_missing_ingredients Documentation}
\usage{
find_missing_ingredients(data)
}
\arguments{
\item{data}{A data frame that to be checked for missing values}
}
\value{
A dataframe containing the columns: column_name, list_of_indices, na_count, na_proportion
}
\description{
Example Description: Returns the indices of rows with columns which have missing values,
as well as how many missing values they have
}
\examples{
tb = tibble("letters" = c("a","b","c"),"numbers" = c(1,2,3))
find_missing_ingredients(tb)
}
